new: newer files are not overwritten by default

+++ src/main.py
@@ -14,6 +14,7 @@ distribute files to all local repos
 import hashlib
 import shutil
 import sys
+from argparse import ArgumentParser
 from pathlib import Path
 
 from utils.file import (
@@ -27,7 +28,12 @@ from utils.trace import Trace
 
 SOURCE_PATH = BASE_PATH
 
-def main() -> None:
+def format_singular_plural(value: int, text: str) -> str:
+    if value == 1:
+        return f"{value} {text}"
+    return f"{value} {text}s"
+
+def main(force: bool = False) -> None:
     repos = Prefs.get("repos")
     Trace.action( f"check {len(repos)} repos" )
 
@@ -47,11 +53,11 @@ def main() -> None:
         for action_type in ["mandatory", "optional", "new"]:
 
             for file in Prefs.get(f"actions.copy.{action_type}.common") or []:
-                modified_files += copy_file_special( SOURCE_PATH, dest, repo["name"], file, action_type )
+                modified_files += copy_file_special( SOURCE_PATH, dest, repo["name"], file, action_type, force=force)
 
             if repo["lib"]:
                 for file in Prefs.get(f"actions.copy.{action_type}.lib") or []:
-                    modified_files += copy_file_special( SOURCE_PATH, dest, repo["name"], file, action_type )
+                    modified_files += copy_file_special( SOURCE_PATH, dest, repo["name"], file, action_type, force=force)
 
         # delete
 
@@ -63,14 +69,14 @@ def main() -> None:
             modified_files_all += modified_files
             modified_repos += 1
 
-    Trace.result( f"{modified_repos} repos, {modified_files_all} files modified" )
+    Trace.result( f"copy {format_singular_plural(modified_files_all, "file")} to {format_singular_plural(modified_repos, "repo")}")
 
 # copy file:
 #  - mandatory -> copy/overwite files
 #  - optional  -> copy/overwrite files, if there is a destination file
 #  - new       -> copy file, if there is NO destination file
 
-def copy_file_special( source: Path, dest: Path, name: str, filepath: Path, action_type: str ) -> int:
+def copy_file_special( source: Path, dest: Path, name: str, filepath: Path, action_type: str, force: bool = False ) -> int:
     src = source / filepath
     dst = dest / filepath
 
@@ -89,15 +95,26 @@ def copy_file_special( source: Path, dest: Path, name: str, filepath: Path, acti
             text = file.read()
         source_md5 = hashlib.md5(text).hexdigest()
 
+        src_timestamp = src.stat().st_mtime
+
+        if src.exists():
+            dst_timestamp = dst.stat().st_mtime
+        else:
+            dst_timestamp = 0
+
         with Path.open(dst, "rb") as file:
             text = file.read()
         dest_md5 = hashlib.md5(text).hexdigest()
 
-        if (source_md5 != dest_md5 ):
-            shutil.copyfile(src, dst)
-            set_modification_timestamp( dst, get_modification_timestamp(src) )
-            Trace.result( f"copy '{filepath}' => {name}" )
-            return 1
+        if (source_md5 != dest_md5):
+            if dst_timestamp>src_timestamp and not force:
+                Trace.error( f"'{name}/{filepath}' is newer" )
+                return 0
+            else:
+                shutil.copyfile(src, dst)
+                set_modification_timestamp( dst, get_modification_timestamp(src) )
+                Trace.result( f"copy '{filepath}' => {name}" )
+                return 1
 
 
     elif action_type in ("mandatory", "new"):
@@ -120,8 +137,12 @@ if __name__ == "__main__":
     Prefs.load("repos.yaml")
     Prefs.load("actions.yaml")
 
+    parser = ArgumentParser(description="distribute common files to all my repos defined in settings/repos.yaml")
+    parser.add_argument("-f", "--force", action="store_true", help="force overwrite newer files")
+    args = parser.parse_args()
+
     try:
-        main()
+        main(args.force)
     except KeyboardInterrupt:
         Trace.exception("KeyboardInterrupt")
         sys.exit()
